move list helpers out of core
