+++ packages/core/src/helpers/index.ts
@@ -1,3 +1,4 @@
+export * from '../../../extension-list-keymap/src/listHelpers/index.js'
 export * from './combineTransactionSteps.js'
 export * from './createChainableState.js'
 export * from './createDocument.js'
@@ -44,7 +45,6 @@ export * from './isNodeActive.js'
 export * from './isNodeEmpty.js'
 export * from './isNodeSelection.js'
 export * from './isTextSelection.js'
-export * from './listHelpers/index.js'
 export * from './posToDOMRect.js'
 export * from './resolveFocusPosition.js'
 export * from './selectionToInsertionEnd.js'
