+++ packages/extension-list-keymap/src/index.ts
@@ -1,5 +1,6 @@
 import { ListKeymap } from './list-keymap.js'
 
 export * from './list-keymap.js'
+export * as listHelpers from './listHelpers/index.js'
 
 export default ListKeymap
