+++ packages/extension-list-keymap/src/listHelpers/getNextListDepth.ts
@@ -1,6 +1,6 @@
 import { EditorState } from '@tiptap/pm/state'
 
-import { getNodeAtPosition } from '../getNodeAtPosition.js'
+import { getNodeAtPosition } from '../../../core/src/helpers/getNodeAtPosition.js'
 import { findListItemPos } from './findListItemPos.js'
 
 export const getNextListDepth = (typeOrName: string, state: EditorState) => {
