+++ packages/extension-list-keymap/src/listHelpers/handleBackspace.ts
@@ -1,8 +1,8 @@
 import { Node } from '@tiptap/pm/model'
 
-import { Editor } from '../../Editor.js'
-import { isAtStartOfNode } from '../isAtStartOfNode.js'
-import { isNodeActive } from '../isNodeActive.js'
+import { Editor } from '../../../core/src/Editor.js'
+import { isAtStartOfNode } from '../../../core/src/helpers/isAtStartOfNode.js'
+import { isNodeActive } from '../../../core/src/helpers/isNodeActive.js'
 import { findListItemPos } from './findListItemPos.js'
 import { hasListBefore } from './hasListBefore.js'
 import { hasListItemBefore } from './hasListItemBefore.js'
