+++ packages/extension-list-keymap/src/listHelpers/listItemHasSubList.ts
@@ -1,7 +1,7 @@
 import { Node } from '@tiptap/pm/model'
 import { EditorState } from '@tiptap/pm/state'
 
-import { getNodeType } from '../getNodeType.js'
+import { getNodeType } from '../../../core/src/helpers/getNodeType.js'
 
 export const listItemHasSubList = (typeOrName: string, state: EditorState, node?: Node) => {
   if (!node) {
