+++ packages/extension-list-keymap/src/listHelpers/nextListIsHigher.ts
